+++ cmd/collector/main.go
@@ -106,7 +106,7 @@ func main() {
 				Aggregator:     aggregator,
 				HealthCheck:    svc.HC(),
 			})
-			collectorOpts := new(app.CollectorOptions).InitFromViper(v)
+			collectorOpts := new(app.CollectorOptions).InitFromViper(v, logger)
 			if err := c.Start(collectorOpts); err != nil {
 				logger.Fatal("Failed to start collector", zap.Error(err))
 			}